Moved useModalHandler to components package.
