+++ packages/components/src/index.ts
@@ -17,3 +17,4 @@ export * from './Spinner';
 export * from './Switch';
 export * from './theme';
 export * from './TooltipTrigger';
+export * from './useModalHandler';
